Adjust folder structure
