+++ source/Calamari/Kubernetes/KubernetesContextScriptWrapper.cs
@@ -12,7 +12,7 @@ using Calamari.Common.Plumbing.Logging;
 using Calamari.Common.Plumbing.Variables;
 using Calamari.Kubernetes.Integration;
 
-namespace Calamari.Kubernetes.Context
+namespace Calamari.Kubernetes
 {
     public class KubernetesContextScriptWrapper : IScriptWrapper
     {
