+++ source/Calamari/Kubernetes/SetupKubectlAuthentication.cs
@@ -7,9 +7,10 @@ using Calamari.Common.Plumbing.FileSystem;
 using Calamari.Common.Plumbing.Logging;
 using Calamari.Common.Plumbing.Proxies;
 using Calamari.Common.Plumbing.Variables;
+using Calamari.Kubernetes.ContextProviders;
 using Calamari.Kubernetes.Integration;
 
-namespace Calamari.Kubernetes.Context
+namespace Calamari.Kubernetes
 {
     public class SetupKubectlAuthentication
     {
